extract notifications in another package, add model Notification
