+++ app/src/main/java/org/stepic/droid/core/StepicCoreComponent.java
@@ -18,7 +18,7 @@ import org.stepic.droid.receivers.InternetConnectionEnabledReceiver;
 import org.stepic.droid.services.CancelLoadingService;
 import org.stepic.droid.services.DeleteService;
 import org.stepic.droid.services.LoadService;
-import org.stepic.droid.services.RegistrationIntentService;
+import org.stepic.droid.notifications.RegistrationIntentService;
 import org.stepic.droid.services.ViewPusher;
 import org.stepic.droid.store.operations.DatabaseFacade;
 import org.stepic.droid.util.ImageOnDisk;
