+++ app/src/main/java/org/stepic/droid/notifications/RegistrationIntentService.kt
@@ -1,4 +1,4 @@
-package org.stepic.droid.services
+package org.stepic.droid.notifications
 
 import android.app.IntentService
 import android.content.Intent
