+++ app/src/main/java/org/stepic/droid/notifications/StepicInstanceIDListenerService.kt
@@ -1,7 +1,8 @@
-package org.stepic.droid.services
+package org.stepic.droid.notifications
 
 import android.content.Intent
 import com.google.android.gms.iid.InstanceIDListenerService
+import org.stepic.droid.notifications.RegistrationIntentService
 
 class StepicInstanceIDListenerService : InstanceIDListenerService() {
     override fun onTokenRefresh() {
