+++ app/src/main/java/org/stepic/droid/view/activities/MainFeedActivity.java
@@ -29,7 +29,7 @@ import org.stepic.droid.events.profile.ProfileCanBeShownEvent;
 import org.stepic.droid.model.EmailAddress;
 import org.stepic.droid.model.Profile;
 import org.stepic.droid.preferences.SharedPreferenceHelper;
-import org.stepic.droid.services.RegistrationIntentService;
+import org.stepic.droid.notifications.RegistrationIntentService;
 import org.stepic.droid.util.AppConstants;
 import org.stepic.droid.view.dialogs.LogoutAreYouSureDialog;
 import org.stepic.droid.view.fragments.DownloadsFragment;
